pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paperstack"
version = "0.1.0"
description = "Academic paper management CLI with web-based PDF viewer and agentic search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Paperstack Authors" }
]
keywords = ["academic", "papers", "research", "cli", "pdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "typer>=0.9.0",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "flask>=2.0.0",
    "flask-cors>=4.0.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "anthropic>=0.18.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "pymupdf>=1.23.0",
    "arxiv>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
paperstack = "paperstack.cli.app:main"

[project.urls]
Homepage = "https://github.com/paperstack/paperstack"
Documentation = "https://github.com/paperstack/paperstack#readme"
Repository = "https://github.com/paperstack/paperstack"

[tool.hatch.build.targets.wheel]
packages = ["src/paperstack"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
